Move the extraction of retrievables into the Retrievables class.
